+++ utils.py
@@ -2,7 +2,7 @@ from PySimpleGUI.PySimpleGUI import Element
 
 def makeArrowKeysWork(elem: Element):
     row = 1
-    if len(elem.Widget.get_children()) > 0:
+    if elem != None and len(elem.Widget.get_children()) > 0:
         selection = elem.Widget.selection()
         if len(selection) > 0:
             row = selection[-1]
@@ -11,4 +11,5 @@ def makeArrowKeysWork(elem: Element):
             elem.Widget.focus(row)  # move focus
             elem.Widget.see(row)  # scroll to show it
         finally:
+            elem.set_focus(True)
             print('could not select items')