+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Adds back/forward navigation with left/right key for the search window in an
   `Undo`/`Redo` manner.
 * Adds volume change with `Shift+Right`/Up and `Shift+Left`/Down
+* Adds status bar, that shows also error for a few seconds, if they occur
+* Adds changing the source with `CMD+K` or `Shift+K`
 
 ## Version 0.3